(PC-2288) : resolve alembic conflict due to multiple heads
